correcciones listas

+++ slim/src/Modules/CourtsModule.php
@@ -31,8 +31,8 @@ final class CourtsModule{
                 ->withStatus(401);
         }
 
-        $body = $req->getBody();
-        $data = json_decode($body, true); // true para que devuelva array asociativo
+        // ahora con body parsing middleware
+        $data = $req->getParsedBody();
         
         $name = $data['name'] ?? '';
         $description = $data['description'] ?? null; // puede ser nulo
@@ -42,30 +42,42 @@ final class CourtsModule{
             return $res->withHeader('Content-Type', 'application/json')->withStatus(400);
         }
 
-        $db = \DB::getConnection();
+        try {
+            $db = \DB::getConnection();
 
-        //si el token es valido, refresco su expiración
-        \Authentication::refreshToken($db, $auth['id'], 300);
+            //si el token es valido, refresco su expiración
+            \Authentication::refreshToken($db, $auth['id'], 300);
 
-        //  Verifico si ya existe una cancha con ese nombre
-        $stmt = $db->prepare("SELECT * FROM courts WHERE name = ?");
-        $stmt->execute([$name]);
-        $row = $stmt->fetch(PDO::FETCH_ASSOC);
+            //  Verifico si ya existe una cancha con ese nombre
+            $stmt = $db->prepare("SELECT * FROM courts WHERE name = ?");
+            $stmt->execute([$name]);
+            $row = $stmt->fetch(PDO::FETCH_ASSOC);
 
-        if (!$row) { // si no existe
-            // Inserto la nueva cancha
-            $insert = $db->prepare("INSERT INTO courts (name, description) VALUES (?, ?)");
-            $insert->execute([$name, $description]);
+            if (!$row) { // si no existe
+                // Inserto la nueva cancha
+                $insert = $db->prepare("INSERT INTO courts (name, description) VALUES (?, ?)");
+                $insert->execute([$name, $description]);
 
-            //  Devuelvo la respuesta
-            $res->getBody()->write(json_encode(['success' => 'Cancha creada correctamente']));
-            return $res->withHeader('Content-Type', 'application/json')->withStatus(201);
-        
-        } else {
-        
-            $res->getBody()->write(json_encode(['error' => 'Ya existe una cancha con ese nombre']));
-            return $res->withHeader('Content-Type', 'application/json')->withStatus(400);
+                //  Devuelvo la respuesta
+                $res->getBody()->write(json_encode(['success' => 'Cancha creada correctamente']));
+                return $res->withHeader('Content-Type', 'application/json')->withStatus(201);
+            
+            } else {
+            
+                $res->getBody()->write(json_encode(['error' => 'Ya existe una cancha con ese nombre']));
+                return $res->withHeader('Content-Type', 'application/json')->withStatus(400);
+            }
+
+
+        } catch (\Throwable $e) {
+        //  \Throwable es la interfaz base de todo lo que puede ser lanzado con throw y capturado con catch.
+            //engloba tanto a las excepciones (Exception) como a los errores fatales (Error).
+
+          error_log($e);
+          $res->getBody()->write(json_encode(['error' => 'Error interno']));
+          return $res->withHeader('Content-Type','application/json; charset=utf-8')->withStatus(500);
         }
+        
     }
 
     // PUT /court/{id} -----------------------------------------------------------------------------------------------
@@ -87,63 +99,71 @@ final class CourtsModule{
                 ->withStatus(401);
         }
 
-        $db = \DB::getConnection();
+        try{
 
-        //si el token es valido, refresco su expiración
-        \Authentication::refreshToken($db, $auth['id'], 300);
+            $db = \DB::getConnection();
 
-        //  Verifico si ya existe una cancha con ese id
-        $stmt = $db->prepare("SELECT * FROM courts WHERE id = ?");
-        $stmt->execute([$id]);
-        $row = $stmt->fetch(PDO::FETCH_ASSOC);
+            //si el token es valido, refresco su expiración
+            \Authentication::refreshToken($db, $auth['id'], 300);
+
+            //  Verifico si ya existe una cancha con ese id
+            $stmt = $db->prepare("SELECT * FROM courts WHERE id = ?");
+            $stmt->execute([$id]);
+            $row = $stmt->fetch(PDO::FETCH_ASSOC);
 
 
-        if ($row) { // si existe
-            
-            $body = $req->getBody();
-            $data = json_decode($body, true); // true para que devuelva array asociativo
-            
-            //tomo los valores a modificar 
-            //$data = $req->getParsedBody(); 
-            $name = $data['name'] ?? '';
-            $description = $data['description'] ?? null; // puede ser nulo
+            if ($row) { // si existe
 
-            // no hay campos por modificar 
-            if (empty($name) && empty($description)) {
-                $res->getBody()->write(json_encode(['error' => 'Debe enviar al menos un campo para modificar']));
-                return $res->withHeader('Content-Type', 'application/json')->withStatus(400);
-            }
+                // ahora con body parsing middleware
+                $data = $req->getParsedBody();
 
-            if (empty($name) || empty($description)){//si solo tengo que modificar alguno de los campos
-
-                if (empty($name) && !empty($description)){ //solo midifico descripcion
-                    
-                    $insert = $db->prepare("UPDATE courts SET description = ? WHERE id = ?");
-                    $insert->execute([$description, $id]);
-                    $row = $insert->fetch(PDO::FETCH_ASSOC);
                 
-                }else{//solo midifico el nombre
+                $name = $data['name'] ?? '';
+                $description = $data['description'] ?? null; // puede ser nulo
+
+                // no hay campos por modificar 
+                if (empty($name) && empty($description)) {
+                    $res->getBody()->write(json_encode(['error' => 'Debe enviar al menos un campo para modificar']));
+                    return $res->withHeader('Content-Type', 'application/json')->withStatus(400);
+                }
+
+                if (empty($name) || empty($description)){//si solo tengo que modificar alguno de los campos
+
+                    if (empty($name) && !empty($description)){ //solo midifico descripcion
+
+                        $insert = $db->prepare("UPDATE courts SET description = ? WHERE id = ?");
+                        $insert->execute([$description, $id]);
+                        $row = $insert->fetch(PDO::FETCH_ASSOC);
                     
-                    $insert = $db->prepare("UPDATE courts SET name = ? WHERE id = ?");
-                    $insert->execute([$name, $id]);
+                    }else{//solo midifico el nombre
+
+                        $insert = $db->prepare("UPDATE courts SET name = ? WHERE id = ?");
+                        $insert->execute([$name, $id]);
+                        $row = $insert->fetch(PDO::FETCH_ASSOC);
+                    }
+
+                }else{ //modifico lo dos campos 
+                    $insert = $db->prepare("UPDATE courts SET name = ?, description = ? WHERE id = ?");
+                    $insert->execute([$name, $description, $id]);
                     $row = $insert->fetch(PDO::FETCH_ASSOC);
                 }
 
-            }else{ //modifico lo dos campos 
-                $insert = $db->prepare("UPDATE courts SET name = ?, description = ? WHERE id = ?");
-                $insert->execute([$name, $description, $id]);
-                $row = $insert->fetch(PDO::FETCH_ASSOC);
+                //  Devuelvo la respuesta
+                $res->getBody()->write(json_encode(['success' => 'Cancha modificada correctamente']));
+                return $res->withHeader('Content-Type', 'application/json')->withStatus(201);
+            
+            } else { //si no existe
+            
+                $res->getBody()->write(json_encode(['error' => 'No existe una cancha con ese ID']));
+                return $res->withHeader('Content-Type', 'application/json')->withStatus(400);
             }
-
-            //  Devuelvo la respuesta
-            $res->getBody()->write(json_encode(['success' => 'Cancha modificada correctamente']));
-            return $res->withHeader('Content-Type', 'application/json')->withStatus(201);
-        
-        } else { //si no existe
-        
-            $res->getBody()->write(json_encode(['error' => 'No existe una cancha con ese ID']));
-            return $res->withHeader('Content-Type', 'application/json')->withStatus(400);
-        }
+        } catch (\Throwable $e) {
+            //  \Throwable es la interfaz base de todo lo que puede ser lanzado con throw y capturado con catch.
+            //engloba tanto a las excepciones (Exception) como a los errores fatales (Error).   
+            error_log($e);  
+            $res->getBody()->write(json_encode(['error' => 'Error interno']));
+            return $res->withHeader('Content-Type','application/json; charset=utf-8')->withStatus(500);
+        }       
 
     }
 
@@ -169,31 +189,50 @@ final class CourtsModule{
                 ->withStatus(401);
         }
         
-        $db = \DB::getConnection();
+        try{
+            $db = \DB::getConnection();
 
-        //si el token es valido, refresco su expiración
-        \Authentication::refreshToken($db, $auth['id'], 300);
-        
-        //busco si tiene reservas
-        $stmt = $db->prepare("SELECT * FROM bookings WHERE court_id = ? ");
-        $stmt->execute([$id]);//retorna F o V
-        $row = $stmt->fetchAll(PDO::FETCH_ASSOC);
+            //si el token es valido, refresco su expiración
+            \Authentication::refreshToken($db, $auth['id'], 300);
 
-        //si res es vacio no hay reservas
-        if (empty($row)){ //si lo elim devuelve v o f
+            //busco si existe la cancha
+            $stmt = $db->prepare("SELECT * FROM courts WHERE id = ?");
+            $stmt->execute([$id]);//retorna F o V
+            $row = $stmt->fetch(PDO::FETCH_ASSOC);
+            if (!($row)){ //si no existe la cancha
+                $res->getBody()->write(json_encode(['error' => 'No existe la cancha']));
+                return $res->withHeader('Content-Type', 'application/json')->withStatus(404);
+            }
 
-            $elim = $db->prepare ("DELETE FROM courts WHERE id = ?");
-            $elim->execute([$id]); //retorna F o V
 
-            $res->getBody()->write(json_encode(['ok' => true, 'deleted_id' => $id]));
-            return $res->withHeader('Content-Type','application/json; charset=utf-8');
-            
 
-        }else {
-            //sino lo pudo elim devuelve las resevas que tiene
-            $res->getBody()->write(json_encode(['error' => 'La cancha tiene reservas']));
-            return $res->withHeader('Content-Type', 'application/json')->withStatus(400);
-        }
+            //busco si tiene reservas
+            $stmt = $db->prepare("SELECT * FROM bookings WHERE court_id = ? ");
+            $stmt->execute([$id]);//retorna F o V
+            $row = $stmt->fetchAll(PDO::FETCH_ASSOC);
+
+            //si res es vacio no hay reservas
+            if (empty($row)){ //si lo elim devuelve v o f
+
+                $elim = $db->prepare ("DELETE FROM courts WHERE id = ?");
+                $elim->execute([$id]); //retorna F o V
+
+                $res->getBody()->write(json_encode(['ok' => true, 'deleted_id' => $id]));
+                return $res->withHeader('Content-Type','application/json; charset=utf-8');
+
+
+            }else {
+                //sino lo pudo elim devuelve las resevas que tiene
+                $res->getBody()->write(json_encode(['error' => 'La cancha tiene reservas']));
+                return $res->withHeader('Content-Type', 'application/json')->withStatus(400);
+            }
+        } catch (\Throwable $e) {
+            //  \Throwable es la interfaz base de todo lo que puede ser lanzado con throw
+            // y capturado con catch. engloba tanto a las excepciones (Exception) como a los errores fatales (Error).
+            error_log($e);
+            $res->getBody()->write(json_encode(['error' => 'Error interno']));
+            return $res->withHeader('Content-Type','application/json; charset=utf-8')->withStatus(500);
+        }  
     }
 
 
@@ -245,7 +284,7 @@ final class CourtsModule{
             return $res->withHeader('Content-Type', 'application/json')->withStatus(500);
         }
     
-        }
+    }
         
 
 }
